Devin2.0-main/requirements.txt
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
astropy==5.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
barfi==0.7.0
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
e2b==0.14.14
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.13.4
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
google_search_results==2.4.2
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonrpcclient==4.0.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.45
langchain-experimental==0.0.57
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langgraph==0.0.26
langsmith==0.1.49
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.7
mdurl==0.1.2
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
openai==1.16.2
orjson==3.10.1
packaging==23.2
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
playwright==1.40.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.0.0
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.8.1b0
pyee==11.0.1
pyerfa==2.0.1.4
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.2
ray==2.8.1
referencing==0.34.0
regex==2024.4.16
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
st-chat-message==0.3.8
stack-data==0.6.3
streamlit==1.33.0
streamlit-chat==0.1.1
streamlit-monaco==0.1.3
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
tiktoken==0.5.2
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
websockets==12.0
yarl==1.9.4
zipp==3.18.1
